An alternative multiply semantic.
A quantity with a conversion factor ratio<N,D,E> where N==1 and D==1 is a coherent quantity
A quantity with a conversion factor ratio<N,D,E> where N==1 and D==1 is an incoherent quantity

Compile time multiplication/division by itself on the incoherent quantity ratio cf (cf * cf, cf^n)  may overflow quite soon in N or D  This is not an academic issue.( see example refrerenced below).
Compile time math on the coherent quantity will never overflow in N or D and only ever overflow with q^integer<intmax_t>::max (unlikely!!!)

Here I show a different semantic for multiplication ( qq_multiply: line 378 /units/quantity.h) which uses this semantic and doesnt suffer from compile time overflow
I show an example in action in example/alternative_multiply_semantic.cpp
Note in there, the commented out part
 // auto xxx = x * x * x ; compile fail : integer overflow

I link also an example in my quan libray, which shows raisng to power 201 and returning back to original quantity using this semantic.

The other advantage of the semantic is you can raise quantitties to arbitrary rational powers and roots, also demonstrated in the link (Would require ratio in mpusz units having a rational exponent)

https://github.com/kwikius/quan-trunk/blob/master/quan_matters/examples/high_power_quantities.cpp

It is as alawys hard to get academics to understand these issue, so if you are happy with the current semantic, fair enough.
Note that with current semantic in mpusz/units, it takes three multiplies to get overflow with us foot, so with this semantic the library is useful for toy applications,
but not really for real industrial use.