units::quantity : Use Concept names rather than class-template names for units::Quantity operator function signatures. - The resulting function signature is shorter and easier to read and understand. - The functions are not restricted to quantity<...> type, but can be reused by other models of Quantity. - Derive quantity class template from base class 'quantity_friend_operations' and put operator functions there as friend functions.      Solves problem with uneven weight of arguments and ODR violations etc - Provide a local concept 'll_quantity_concept' representing only units::quantity<...> class template.
examples :
Integrate std::chrono::duration as a model of units::Quantity
The model of units::Quantity now requires get_dimension<Q>::type and get_unit<Q>::type traits rather than peeking the type directly.
In example/chrono_duration_compat.cpp we make chrono::duration a model of units::Quantity.
N.B This is a proof of concept and only the minimal work is done here to get this working for this minimal example.
Upgrading the rest is left as an excercise...