Fix for slow compilation with apple-clang
The new version of the `Representation` concept using
`detail::ScalarRepresentation`, `detail::ComplexRepresentation`,
`detail::VectorRepresentation`, `detail::TensorRepresentation` was
causing some translation units to take two orders of magnitude longer to
compile.  This was observed with Xcode 15.1 and 15.4.

As a workaround use the old implementation of `Representation` when
building with apple-clang.

The new implementation should be tested with Xcode 16 to see if this
workaround needs to be applied to all shipping versions of apple-clang
or <16.

More investigation could be done to try to allow the new constrained
behavior in a more efficient way now that we have isolated the issue.